+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.2.3
+- New:
+  - Between date fields now show start/end date fields
+  
+- Improved:
+  - Some fields broken out into separate components (WIP)
+
 # 0.2.2
 
 - Fixed: